feat(output): new pdfboss-output crate owns layout and text rendering
Moves layout()/segments()/flow() and the extract_text* entry points out of pdfboss-text unchanged; pdfboss-text is spans-only. All consumers migrated; no output change.
