+++ crates/pdfboss-cli/Cargo.toml
@@ -10,7 +10,7 @@ keywords.workspace = true
 
 [dependencies]
 pdfboss-core = { workspace = true }
-pdfboss-text = { workspace = true }
+pdfboss-output = { workspace = true }
 pdfboss-render = { workspace = true }
 pdfboss-aio = { workspace = true, features = ["http"] }
 pdfboss-tui = { workspace = true }
