+++ crates/pdfboss-py/Cargo.toml
@@ -14,7 +14,7 @@ crate-type = ["cdylib"]
 
 [dependencies]
 pdfboss-core = { path = "../pdfboss-core" }
-pdfboss-text = { path = "../pdfboss-text" }
+pdfboss-output = { path = "../pdfboss-output" }
 pdfboss-render = { path = "../pdfboss-render" }
 # `http` is enabled here so every wheel/sdist build ships open_url support.
 pdfboss-aio = { path = "../pdfboss-aio", features = ["http"] }
